Move the custom editor api out of proposed
For #77131
